requirements.txt
alembic==1.4.2
appdirs==1.4.4
arrow==0.16.0
Babel==2.8.0
binaryornot==0.4.4
black==20.8b1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==3.1
decorator==4.4.2
elasticsearch==7.10.1
FormEncode==1.3.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
hupper==1.10.2
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.21
neo4j==4.1.1
nose==1.3.7
paginate==0.5.6
PasteDeploy==2.1.0
pathspec==0.8.0
Pillow==7.2.0
plaster==1.0
plaster-pastedeploy==0.7
poyo==0.5.0
protobuf==3.13.0
pycparser==2.20
Pygments==2.6.1
pyramid==1.10.4
pyramid-authstack==1.0.1
pyramid-debugtoolbar==4.6.1
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pyramid-retry==2.1.1
pyramid-session-redis==1.5.2
pyramid-tm==2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
PyUtilib==6.0.0
redis==3.5.3
regex==2020.7.14
repoze.lru==0.7
requests==2.24.0
rutter==0.3
six==1.15.0
SQLAlchemy==1.3.19
text-unidecode==1.3
timeago==1.0.14
toml==0.10.1
transaction==3.0.0
translationstring==1.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
validators==0.18.1
venusian==3.0.0
waitress==1.4.4
WebHelpers2==2.0
WebOb==1.8.6
zope.deprecation==4.4.0
zope.event==4.4
zope.interface==5.1.0
zope.sqlalchemy==1.3